requirements.txt
# Core
fastapi==0.135.1
uvicorn[standard]==0.41.0
python-dotenv==1.2.2
pydantic==2.12.5
python-multipart==0.0.20

# LLM e Agente
langchain==1.2.12
langchain-anthropic==1.3.4
langchain-openai==1.1.11
langchain-google-genai==4.2.1
langchain-ollama==1.0.1
langgraph==1.1.1
anthropic==0.84.0
google-generativeai==0.8.6
openai==2.26.0

# Voz
faster-whisper==1.2.1
edge-tts==7.2.7
sounddevice==0.5.5
soundfile==0.13.1
pvporcupine==4.0.2

# Memória
chromadb==1.5.5
sentence-transformers==5.2.3
aiosqlite==0.20.0

# Agendamento
apscheduler==3.11.2

# Automação
playwright==1.58.0
ytmusicapi==1.10.2

# Integrações
google-api-python-client==2.192.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
spotipy==2.26.0
python-telegram-bot==22.6

# Notícias
feedparser==6.0.11

# Sistema
psutil==6.1.1

# Logs
loguru==0.7.3

# Web Scraping
beautifulsoup4==4.12.3

# Dev
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
